Simplify rate limit
This was just too confusing so simplify and you pretty much get the same result with only one much more obvious setting.
Also handle the case where the limit was previously very high, so the users spawn count could be big, but then we switch to a much smaller limit and they have to wait a long time for the recharge to happen.