Persist exempted users
I could have persisted these in per player properties but for now I'm sticking with this in memory list to give me the option later to allow an admin to add/remove people from it.
It *might* be faster than querying a per player property, but then I guess that could be cached too.